fix: signMessage test
